refactor : 결제 금액 타입 Long → int 통일 및 OrderService 단위 테스트 추가

+++ Mavis-Api/src/main/java/com/mavis/api/order/service/OrderService.java
@@ -80,7 +80,7 @@ public class OrderService {
     }
 
     @Transactional(readOnly = true)
-    public Order validateOrderForPayment(String orderId, Long amount) {
+    public Order validateOrderForPayment(String orderId, int amount) {
         Order order = orderRepository.findByOrderIdAndIsDeletedFalse(orderId)
                 .orElseThrow(() -> OrderNotFoundException.EXCEPTION);
