+++ Mavis-Infrastructure/src/main/java/com/mavis/infrastructure/outer/api/tosspayments/dto/TossConfirmRequest.java
@@ -3,9 +3,9 @@ package com.mavis.infrastructure.outer.api.tosspayments.dto;
 public record TossConfirmRequest(
         String paymentKey,
         String orderId,
-        Long amount
+        int amount
 ) {
-    public static TossConfirmRequest of(String paymentKey, String tossOrderId, Long amount) {
+    public static TossConfirmRequest of(String paymentKey, String tossOrderId, int amount) {
         return new TossConfirmRequest(paymentKey, tossOrderId, amount);
     }
 }
